Models/CV Parser/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
pdfplumber==0.10.2
google-generativeai==0.3.2
werkzeug==2.3.7